Fix amounts overflow (#221)
* WIP: Migrate all the amounts to BigDecimal

* Migrate amounts to BigDecimal, but keep Account for now

* Restore DB schema for version 9, update DB version to 10

* Make all amounts in Account use BigDecimal

* Fix showing multiplied token transfer amount on the send receipt screen

* Get rid of CurrencyUtil.formatGTU method with value of type long

* Fix unit tests, add big value tests for CurrencyUtil

* Remove useless BigDecimal scaling extension

* Fix instrumeted tests, add test for account balance

* Replace BigDecimal with BigInteger to simplify things

* Omit useless `this` in the BigInteger extension


---------

+++ app/src/main/java/com/concordium/wallet/data/backend/OfflineMockInterceptor.kt
@@ -202,9 +202,9 @@ class OfflineMockInterceptor : Interceptor {
                 "013c6d2dd67affd6f39b9a7b255d244055b53d68fe8b0add4839a20e911d04cb",
                 timeStamp.toDouble(),
                 "84bf1e2ef8d3af3063cdb681932990f71ddb3949655f55307a266e5d687b414f",
-                -10000,
-                59,
-                -10059,
+                (-10000).toBigInteger(),
+                59.toBigInteger(),
+                (-10059).toBigInteger(),
                 59,
                 details,
                 null
