+++ app/src/main/java/com/concordium/wallet/ui/account/accountsoverview/AccountsOverviewFragment.kt
@@ -31,6 +31,7 @@ import com.concordium.wallet.ui.common.delegates.IdentityStatusDelegateImpl
 import com.concordium.wallet.ui.identity.identitiesoverview.IdentitiesOverviewActivity
 import com.concordium.wallet.ui.identity.identityproviderlist.IdentityProviderListActivity
 import com.concordium.wallet.ui.transaction.sendfunds.SendFundsActivity
+import java.math.BigInteger
 
 class AccountsOverviewFragment : BaseFragment(),
     IdentityStatusDelegate by IdentityStatusDelegateImpl(), EarnDelegate by EarnDelegateImpl() {
@@ -486,15 +487,15 @@ class AccountsOverviewFragment : BaseFragment(),
         binding.createAccountButton.visibility = state
     }
 
-    private fun showTotalBalance(totalBalance: Long) {
+    private fun showTotalBalance(totalBalance: BigInteger) {
         binding.totalBalanceTextview.text = CurrencyUtil.formatGTU(totalBalance)
     }
 
-    private fun showDisposalBalance(atDisposal: Long) {
+    private fun showDisposalBalance(atDisposal: BigInteger) {
         binding.accountsOverviewTotalDetailsDisposal.text = CurrencyUtil.formatGTU(atDisposal, true)
     }
 
-    private fun showStakedBalance(totalBalance: Long) {
+    private fun showStakedBalance(totalBalance: BigInteger) {
         binding.accountsOverviewTotalDetailsStaked.text = CurrencyUtil.formatGTU(totalBalance, true)
     }
